Fixed race conditions in LockableWindow and PersistenceWindowPool.
It was possible to close a LockableWindow even though another thread had locked that window and was using it.
Fixing this issue also unvieled another concurrency issue in PersistenceWindowPool where it was possible to close a window without removing it from the pool.
In order to be allowed to close a LockableWindow these conditions must hold:
1. The window must not be in use (locked) by some other thread.
2. The window must not be marked by another thread for intention to use it.
These conditions are encapsulated in the isFree() method.
Unit tests have been added to cover the disired behavior of when it is ok to close a window.